refactor(video-annotation): unify the two TD-version hooks
