+++ app/packages/video-annotation/src/sync/useSyncSidebarFromTemporalOverlays.ts
@@ -2,19 +2,14 @@
  * Copyright 2017-2026, Voxel51, Inc.
  */
 
-import {
-  type Scene2D,
-  TemporalOverlay,
-  UNDEFINED_LIGHTER_SCENE_ID,
-  useLighterEventHandler,
-} from "@fiftyone/lighter";
+import { type Scene2D, TemporalOverlay } from "@fiftyone/lighter";
 import {
   type AnnotationLabel,
   type AnnotationLabelData,
   useModalSample,
 } from "@fiftyone/state";
 import { isFrameInSupport } from "@fiftyone/utilities";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useEffect, useRef } from "react";
 import {
   useGetSidebarLabels,
   useLabelsContext,
@@ -22,6 +17,7 @@ import {
 import useFocus from "../../../core/src/components/Modal/Sidebar/Annotate/useFocus";
 import { frameAt } from "../../../playback/src/lib/playback/utils";
 import { usePlayhead } from "../../../playback/src/lib/playback/use-playback-state";
+import { useTemporalOverlayVersion } from "../hooks/useTemporalOverlayVersion";
 import { getModalSampleFrameRate } from "../utils/modalSample";
 
 const TEMPORAL_DETECTION = "TemporalDetection" as const;
@@ -81,7 +77,7 @@ export const useSyncSidebarFromTemporalOverlays = (
   // reconcile below wouldn't run until the next playhead tick, so a freshly
   // created TD never enters the sidebar. Mirror `FrameLabels`' TD-track
   // invalidation: bump on scene TD overlay add/remove.
-  const sceneTdVersion = useSceneTemporalOverlayVersion(scene);
+  const sceneTdVersion = useTemporalOverlayVersion(scene);
 
   useEffect(() => {
     if (!scene || !canonicalMediaReady || frame === null) {
@@ -157,42 +153,3 @@ export const useSyncSidebarFromTemporalOverlays = (
     updateLabelData,
   ]);
 };
-
-/**
- * A counter that bumps whenever a `TemporalOverlay` is added to, or a `td-`
- * overlay removed from, the scene. Lets the reconcile effect re-run on
- * scene-only TD changes that the sidebar-derived signature can't observe.
- */
-function useSceneTemporalOverlayVersion(scene: Scene2D | null): number {
-  const useLighterEvent = useLighterEventHandler(
-    scene?.getEventChannel() ?? UNDEFINED_LIGHTER_SCENE_ID
-  );
-  const [version, setVersion] = useState(0);
-  const bump = useCallback(() => setVersion((v) => v + 1), []);
-
-  useLighterEvent(
-    "lighter:overlay-added",
-    useCallback(
-      (payload) => {
-        if (payload.overlay instanceof TemporalOverlay) {
-          bump();
-        }
-      },
-      [bump]
-    )
-  );
-
-  useLighterEvent(
-    "lighter:overlay-removed",
-    useCallback(
-      (payload) => {
-        if (payload.id?.startsWith("td-")) {
-          bump();
-        }
-      },
-      [bump]
-    )
-  );
-
-  return version;
-}
